Move caching methods to valdator.py
...so they can be used when the validator is used as a library.
